section 9: split rbformats/mod.rs — extract compress/decompress dispatch
New src/rbformats/compress.rs (658 lines) owns compress_partition,
decompress_to_writer, decompress_partition_to_file, compress_file_to_archive,
and the small shared helpers (write_zeros, write_zeros_with_progress,
is_all_zeros, output_path, file_name, SplitWriter, CHUNK_SIZE) plus their
eight tests. mod.rs re-exports public/pub(crate) items so existing call
sites compile unchanged. mod.rs: 1703 → 1067 lines (-636).
